% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ProcessData.R
\name{BioQCREtoNet}
\alias{BioQCREtoNet}
\title{Compute the Quaternary Product Scoring Statistic for a biological causal network.}
\usage{
BioQCREtoNet(relations, evidence, entities, method = "Quaternary",
                    fc.thresh = 1.3, is.Logfc = TRUE, pval.thresh = 0.01)
}
\arguments{
\item{relations}{A data frame containing pairs of connected entities in a causal network (e.g Protein-Protein interactions),
and the type of causal relation between them. The data frame must have three columns with column names: \emph{srcuid},
\emph{trguid} and \emph{mode} respective of order. \emph{srcuid} stands for source entity, \emph{trguid} stands for target entity and \emph{mode} stands for the
type of relation between \emph{srcuid} and \emph{trguid}. The relation has to be one of \emph{increases}, \emph{decreases} or \emph{regulates}. All three columns must be of type character.}

\item{evidence}{A data frame of entities which are target nodes in the causal network and have new
experimental values (e.g gene expression data). The \emph{evidence} data frame must have three columns \emph{entrez}, \emph{fc} and \emph{pvalue}.
\emph{entrez} denotes the entrez id of a given gene, \emph{fc} denotes
the fold change of a gene, and \emph{pvalue} denotes the p-value. The \emph{entrez} column must be of type
character, and the \emph{fc} and \emph{pvalue} columns must be numeric values.}

\item{entities}{A data frame of mappings for all entities present in data frame \emph{relations}. \emph{entities} must contain
four columns: \emph{uid}, \emph{id}, \emph{symbol} and \emph{type} respective of order. All four columns must
be of type character. \emph{uid} includes every source and target node in the network (i.e \emph{relations}),
\emph{id} is the id of \emph{uid} (e.g entrez id of an mRNA), \emph{symbol} is the symbol of \emph{id} and \emph{type} is the type
of entity of \emph{id} which can be one of mRNA, protein, drug, compound etc. All target nodes must be of type mRNA.}

\item{method}{Choose one of \emph{Quaternary}, \emph{Ternary} or \emph{Enrichment}. Default is \emph{Quaternary}.}

\item{fc.thresh}{Threshold for fold change in \emph{evidence} data frame. Any row in evidence with abosolute value \emph{fc}
smaller than \emph{fc.thresh} will be ignored. Default value is 1.3. If
\emph{is.Logfc = TRUE} then \emph{fc.thresh} is set to \emph{log2(fc.thresh)} internally.}

\item{is.Logfc}{Boolean value to indicate if the fold change is in log scale. Default value is TRUE.}

\item{pval.thresh}{Threshold for p-values in \emph{evidence} data frame. All rows in \emph{evidence} with p-values
       greater than \emph{pval.thresh} will be ingnored. Default value is 0.01.}
}
\value{
This function returns a data frame containing parameters concerning the Quaternary Product
        Scoring Statistic. The p-values of each of the source nodes is also computed, and the data frame
        is in increasing order of p-values of the goodness of fit score for the given source nodes. The column
        names of the data frame are:

\itemize{
\item  \emph{uid} The source node in the network.
\item \emph{name} symbols of the source nodes.
\item \emph{regulation} Direction of change of source node.
\item \emph{correct.pred} Number of correct predictions in \emph{evidence} when compared to predictions made
                    by the network.
\item \emph{incorrect.pred} Number of incorrect predictions in \emph{evidence} when compared to predictions made
                    by the network.
\item \emph{score} Value of the Quaternary Product Scoring Statistic.
\item \emph{total.reachable} Total number of \emph{trguid}s connected to a \emph{srcuid}.
\item \emph{significant.reachable} number of \emph{trguid}s connected to a \emph{srcuid} that are also regulated in \emph{evidence}.
\item \emph{total.ambiguous} Total number of \emph{srcuid} children with relation type \emph{regulates} or children
                             which share both \emph{increase} and \emph{decrease} relation with \emph{srcuid}.
\item \emph{significant.ambiguous} Total number of similar type of relations as with \emph{total.ambiguous} but
                                   with the restriction that the children are regulated in \emph{evidence}.
\item \emph{unknown} All \emph{trguid}s which do not interact with the given \emph{srcuid}.
\item \emph{pvalue} P-value of the score.
}
}
\description{
This function computes the Quaternary Product Scoring Statistic for all source nodes in a
             causal network once new gene expression data is presented. This is a Causal Relation Engine (CRE)
             on the levels of a causal network of predictions and a new set of realizations that
             arise in real world situations.
}
\author{
Carl Tony Fakhry, Ping Chen and Kourosh Zarringhalam
}
\references{
C. T. Fakhry, P. Choudhary, A. Gutteridge, B. Sidders, P. Chen, D. Ziemek, K. Zarringhalam.
            Identifying Transcriptional Regulators in Signed and Unsigned Causal Networks, 2015, In submission.
}

