\name{QuaternaryProd-package}
\alias{QuaternaryProd-package}
%\alias{QuaternaryProd}
\docType{package}
\title{
Quaternary Product Scoring Statistic for Signed Causal Graphs
}
\description{
This package computes the Quaternary Product Scoring Statistic for signed causal graphs.
}
\details{
\tabular{ll}{
Package: \tab QuaternaryProd\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2015-10-22\cr
License: \tab GPL (>= 2)\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
Carl Tony Fakhry, Ping Chen and Kourosh Zarringhalam 

Maintainer: Carl Tony Fakhry <cfakhry@cs.umb.edu>
}
\references{
Fakhry, C. T. et al. An Efficient Algorithm for Computing the Quaternary Product Scoring
Statistic for Signed Causal Graphs with Ambiguities. under review.
}

%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%% ~~ simple examples of the most important functions ~~
%}
