% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessData.R
\name{RunCRE_HSAStringDB}
\alias{RunCRE_HSAStringDB}
\title{This function runs a causal relation engine by computing the Quaternary Dot
Product Scoring Statistic, Ternary Dot Product Scoring Statistic or the Enrichment test over the Homo
Sapien STRINGdb causal network.}
\usage{
RunCRE_HSAStringDB(gene_expression_data, method = "Quaternary", 
                    fc.thresh = 1.3, pval.thresh = 0.05)
}
\arguments{
\item{gene_expression_data}{A data frame for gene expression data. The \emph{gene_expression_data} data frame must have three columns \emph{entrez}, 
\emph{fc} and \emph{pvalue}. \emph{entrez} denotes the entrez id of a given gene, \emph{fc} denotes
the fold change of a gene, and \emph{pvalue} denotes the p-value. The \emph{entrez} column must be of type
integer or character, and the \emph{fc} and \emph{pvalue} columns must be numeric values.}

\item{method}{Choose one of \emph{Quaternary}, \emph{Ternary} or \emph{Enrichment}. Default is \emph{Quaternary}.}

\item{fc.thresh}{Threshold for fold change in \emph{gene_expression_data} data frame. Any row in gene_expression_data with abosolute value of \emph{fc}
smaller than \emph{fc.thresh} will be ignored. Default value is 1.3.}

\item{pval.thresh}{Threshold for p-values in \emph{gene_expression_data} data frame. All rows in \emph{gene_expression_data} with p-values 
greater than \emph{pval.thresh} will be ingnored. Default value is 0.05.}
}
\value{
This function returns a data frame containing parameters concerning the method used. The p-values of each
        of the source nodes is also computed, and the data frame
        is in increasing order of p-values of the goodness of fit score for the given source nodes. The column
        names of the data frame are:
        
\itemize{        
\item  \emph{uid} The source node in the STRINGdb network.
\item \emph{name} Symbol of the source node. 
\item \emph{regulation} Direction of change of source node.
\item \emph{correct.pred} Number of correct predictions in \emph{gene_expression_data} when compared to predictions made
                    by the network.
\item \emph{incorrect.pred} Number of incorrect predictions in \emph{gene_expression_data} when compared to predictions made
                    by the network.
\item \emph{score} The number of correct predictions minus the number of incorrect predictions. 
\item \emph{total.reachable} Total Number of children of the given source node.
\item \emph{significant.reachable} Number of children of the given source node that are also present 
                                   in \emph{gene_expression_data}.
\item \emph{total.ambiguous} Total number of children of the given source node which are regulated by the given source node without
                             knowing the direction of regulation.
\item \emph{significant.ambiguous} Total number of children of the given source node which are regulated by the given source node without
                             knowing the direction of regulation and are also present in \emph{gene_expression_data}.  
\item \emph{unknown} Number of target nodes in the STRINGdb causal network which do not interact with the given source node.
\item \emph{pvalue} P-value of the score computed according to the selected method.
}
}
\description{
This function runs a causal relation engine by computing the Quaternary Dot
             Product Scoring Statistic, Ternary Dot Product Scoring Statistic or the Enrichment test over the Homo
             Sapien STRINGdb causal network.
}
\examples{

# Get gene expression data
gene_expression_data1 <- system.file("extdata", "e2f3_sig.txt", package = "QuaternaryProd")
gene_expression_data1 <- read.table(gene_expression_data1, sep = "\\t", header = TRUE
                                                  , stringsAsFactors = FALSE)

# Remove duplicated entrez ids in gene_expression_data1 and rename column names appropriately
gene_expression_data1 <- gene_expression_data1[!duplicated(gene_expression_data1$entrez),]
names(gene_expression_data1) <- c("entrez", "pvalue", "fc")
          
# Compute the statistic for each source node in the String network
results <- RunCRE_HSAStringDB(gene_expression_data1)

}
\references{
Carl Tony Fakhry, Parul Choudhary, Alex Gutteridge, Ben Sidders, Ping Chen, Daniel Ziemek, and
            Kourosh Zarringhalam. Interpreting transcriptional changes using causal graphs: new methods and
            their practical utility on public networks. BMC Bioinformatics, 17:318, 2016. ISSN 1471-2105.
            doi: 10.1186/s12859-016-1181-8.
           
            Franceschini, A (2013). STRING v9.1: protein-protein interaction networks, with increased coverage 
            and integration. In:'Nucleic Acids Res. 2013 Jan;41(Database issue):D808-15. doi: 10.1093/nar/gks1094. 
            Epub 2012 Nov 29'.
}
\author{
Carl Tony Fakhry, Ping Chen and Kourosh Zarringhalam
}
